Split up search.h a little
